+++ src/traits/THasName.php
@@ -1,9 +1,9 @@
 <?php
 /**
+ * @link https://cms.skeeks.com/
+ * @copyright Copyright (c) 2010 SkeekS
+ * @license https://cms.skeeks.com/license/
  * @author Semenov Alexander <semenov@skeeks.com>
- * @link https://skeeks.com/
- * @copyright 2010 SkeekS
- * @date 05.03.2017
  */
 
 namespace skeeks\cms\traits;
@@ -11,13 +11,11 @@ namespace skeeks\cms\traits;
 use yii\helpers\ArrayHelper;
 
 /**
- * @deprecated
- *
  * @property $name;
- * @property $icon;
- * @property $image;
+ *
+ * @author Semenov Alexander <semenov@skeeks.com>
  */
-trait THasInfo
+trait THasName
 {
     /**
      * @var string
@@ -48,55 +46,7 @@ trait THasInfo
         } else if (is_string($name)) {
             $this->_name = $name;
         }
-        
-        return $this;
-    }
 
-
-    /**
-     * @var string
-     */
-    protected $_icon = '';
-
-    /**
-     * @return string
-     */
-    public function getIcon()
-    {
-        return $this->_icon;
-    }
-
-    /**
-     * @param $icon
-     * @return $this
-     */
-    public function setIcon($icon)
-    {
-        $this->_icon = $icon;
-        return $this;
-    }
-
-
-    /**
-     * @var string
-     */
-    protected $_image = '';
-
-    /**
-     * @return string
-     */
-    public function getImage()
-    {
-        return $this->_image;
-    }
-
-    /**
-     * @param $image
-     * @return $this
-     */
-    public function setImage($image)
-    {
-        $this->_image = $image;
         return $this;
     }
 }